# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ToNote App OpenApi API Documentation",
        "description": "ToNote App Using L5 Swagger OpenApi description",
        "contact": {
            "email": "schneidershades@gmail.com"
        },
        "version": "3.0"
    },
    "servers": [
        {
            "url": "http://getwordprime.herokuapp.com/",
            "description": "Staging API server"
        }
    ],
    "paths": {
        "/api/v1/admin/agencies": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get Agencys",
                "description": "Get Agencys",
                "operationId": "allAdminAgencies",
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Post agencies",
                "description": "Post agencies",
                "operationId": "postAdminAgencies",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgencyCreateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/admin/agencies/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Show agency",
                "description": "Show agency",
                "operationId": "showAgencys",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Agencys ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update agency",
                "description": "Update agency",
                "operationId": "updateAdminAgencys",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "agency ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgencyUpdateFormRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete agency",
                "description": "Delete agency",
                "operationId": "deleteAdminAgencys",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Agencys ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful signin",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AgencyCreateFormRequest": {
                "title": "Agency Create Form Request Fields",
                "description": "Agency Create request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "agency name",
                        "description": "name of the agency",
                        "type": "string",
                        "example": "Info Limited"
                    },
                    "email": {
                        "title": "agency email",
                        "description": "email of the agency",
                        "type": "string",
                        "example": "info@client.com"
                    },
                    "campaigns": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "campaign_id": {
                                    "type": "integer",
                                    "example": "1"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "AgencyUpdateFormRequest": {
                "title": "Agency Update Form Request Fields",
                "description": "Agency Update request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "agency name",
                        "description": "name of the agency",
                        "type": "string",
                        "example": "Info Limited"
                    },
                    "email": {
                        "title": "agency email",
                        "description": "email of the agency",
                        "type": "string",
                        "example": "info@client.com"
                    },
                    "campaigns": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "campaign_id": {
                                    "type": "integer",
                                    "example": "1"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}